Links and redirects

Closes #97 and #99

See merge request team-project-h/2022/sh24/sh24-main!53

+++ django-backend/allocationapp/views.py
@@ -25,7 +25,10 @@ def index(request):
     if is_manager(request.user):
         return redirect(reverse('allocationapp:manager_view_teams'))
     elif is_grad(request.user):
-        return redirect(reverse('allocationapp:cast_votes'))
+        if grad_has_already_voted(request.user):
+            return redirect(reverse('allocationapp:vote_submitted'))
+        else:
+            return redirect(reverse('allocationapp:cast_votes'))
     elif is_admin(request.user):
         return redirect(reverse('allocationapp:portal'))
     else:
@@ -56,11 +59,13 @@ def cast_votes(request):
 
         return redirect(reverse('allocationapp:vote_submitted'))
     else:
-        context_dict = {}
+        context_dict = {
+            'allocation_run': allocation_run(),
+        }
 
         # Get all the votes cast by the graduate
-        votes = Preference.objects.filter(graduate=Graduate.objects.get(user=CustomUser.objects.get(id=request.user.id)))
-        if len(votes) > 0:
+        votes = Preference.objects.filter(graduate=Graduate.objects.get(user=CustomUser.objects.get(id=current_user.id)))
+        if grad_has_already_voted(current_user):
             # Grad has previously cast their votes.
             team_id_to_votes = {}
             for vote in votes:
@@ -103,6 +108,7 @@ def result_page(request):
         'assigned_team': current_user.assigned_team,
         'assigned_team_members': assigned_team_members,
         'current_user_id': request.user.id,
+        'allocation_run': allocation_run(),
     }
 
     return render(request, 'allocationapp/result_page.html', context=context_dict)
@@ -234,6 +240,9 @@ def add_new_technology(request, team_id, tech_name):
 @login_required
 @user_passes_test(is_admin, login_url='/allocation/')
 def upload_file(request):
+    if allocation_run():
+        return redirect(reverse('allocationapp:portal'))
+
     if request.method == 'POST':
         form = CSVForm(request.POST, request.FILES)
         
@@ -250,12 +259,15 @@ def upload_file(request):
     else:
         form = CSVForm()
     all_csv = UserCSV.objects.all()
-    return render(request, 'allocationapp/upload.html', {'form': form, 'all_csv': all_csv, 'populated': False})
+    return render(request, 'allocationapp/upload.html', {'form': form, 'all_csv': all_csv, 'populated': False, 'allocation_ran': allocation_run(),})
 
 
 @login_required
 @user_passes_test(is_admin, login_url='/allocation/')
 def populate_db(request):
+    if allocation_run():
+        return redirect(reverse('allocationapp:portal'))
+
     reset_graduates_managers()
     try:
         csv_file = UserCSV.objects.get(pk=1).csv_file
@@ -333,6 +345,9 @@ def populate_db(request):
 @login_required
 @user_passes_test(is_admin, login_url='/allocation/')
 def team_upload_file(request):
+    if allocation_run():
+        return redirect(reverse('allocationapp:portal'))
+    
     if request.method == 'POST':
         form = CSVForm(request.POST, request.FILES)
         if form.is_valid() and len(request.FILES['csv_file']) > 0:
@@ -348,12 +363,15 @@ def team_upload_file(request):
         form = CSVForm()
 
     all_csv = TeamCSV.objects.all()
-    return render(request, 'allocationapp/team_upload.html', {'form': form, 'all_csv': all_csv, 'populated': False})
+    return render(request, 'allocationapp/team_upload.html', {'form': form, 'all_csv': all_csv, 'populated': False, 'allocation_ran': allocation_run()})
 
 
 @login_required
 @user_passes_test(is_admin, login_url='/allocation/')
 def team_populate_db(request):
+    if allocation_run():
+        return redirect(reverse('allocationapp:portal'))
+
     reset_teams()
     try:
         csv_file = TeamCSV.objects.get(pk=1).csv_file
@@ -431,6 +449,9 @@ def team_populate_db(request):
 @login_required
 @user_passes_test(is_admin, login_url='/allocation/')
 def reset_teams_view(request):
+    if allocation_run():
+        return redirect(reverse('allocationapp:portal'))
+
     try:
         reset_teams()
         TeamCSV.objects.all().delete()
@@ -443,6 +464,9 @@ def reset_teams_view(request):
 @login_required
 @user_passes_test(is_admin, login_url='/allocation/')
 def reset_graduates_managers_view(request):
+    if allocation_run():
+        return redirect(reverse('allocationapp:portal'))
+
     try:
         reset_graduates_managers()
         UserCSV.objects.all().delete()
@@ -473,6 +497,9 @@ def get_allocation(request):
 @login_required
 @user_passes_test(is_admin, login_url='/allocation/')
 def create_new_team(request):
+    if allocation_run():
+        return redirect(reverse('allocationapp:portal'))
+
     if request.method == 'POST':
         name = request.POST['group_name']
         manager = request.POST['group_manager']
@@ -518,13 +545,16 @@ def create_new_team(request):
     technologies = Technology.objects.all()
     managers = Manager.objects.all()
     context_dict = {'managers': managers, 'departments': departments, 'skills': skills, 'technologies': technologies,
-                    'count': range(0, 200)}
+                    'count': range(0, 200), 'allocation_ran': allocation_run(),}
     return render(request, 'allocationapp/create_new_team.html', context=context_dict)
 
 
 @login_required
 @user_passes_test(is_admin, login_url='/allocation/')
 def create_new_grad(request):
+    if allocation_run():
+        return redirect(reverse('allocationapp:portal'))
+
     if request.method == 'POST':
         first_name = request.POST['first_name']
         last_name = request.POST['last_name']
@@ -551,7 +581,7 @@ def create_new_grad(request):
         messages.success(request, f'Added {first_name} {last_name} to the database!')
         return redirect(reverse('allocationapp:create_new_grad'))
 
-    return render(request, 'allocationapp/create_new_graduate.html')
+    return render(request, 'allocationapp/create_new_graduate.html', context={'allocation_ran': allocation_run()})
 
 
 @login_required
